NotePadDev
I have added a SaveFileDialog when the user clicks "Save as". I have to add the funcions to save the file (when it already exists) and to open a file to create a Document instance (for this I have to add a constructor to the Document class that gets the text to initialize the object from a file). Use the same structure used in the java programs